refactor : app 내부에서 game controller를 호출한 후 game에서 view를 호출하는 방식으로 변경
